Propagate task cancellation
Cancel TaskCompletionSource when supplied CT is cancelled (ie download request is cancelled by Uno Image). This is because UIL seems to internally cancel previous downloads if a new URI is requested for the same ImageView (without raising IImageListener callbacks), leading to tasks that never return, but are still awaited for subsequent requests to the same URI thanks to the concurrency-management logic.